+++ packages/sdk-ink/src/index.ts
@@ -1,4 +1,4 @@
 export { createInkSdk } from "./ink-sdk"
 export { createReviveSdk } from "./revive-sdk"
 export type * from "./descriptor-types"
-export { ss58ToEthereum } from "./util"
+export { ss58ToEthereum, getDeploymentAddressWithNonce } from "./util"