+++ packages/sdk-ink/src/util.ts
@@ -1,6 +1,15 @@
+import { RLP } from "@ethereumjs/rlp"
 import { keccak_256 } from "@noble/hashes/sha3"
-import { AccountId, Binary, Enum, SS58String } from "polkadot-api"
-import { U256 } from "./descriptor-types"
+import {
+  AccountId,
+  Binary,
+  Enum,
+  FixedSizeBinary,
+  HexString,
+  SS58String,
+} from "polkadot-api"
+import { mergeUint8 } from "polkadot-api/utils"
+import { ReviveAddress, U256 } from "./descriptor-types"
 
 export const getSignedStorage = (
   depositResponse: Enum<{
@@ -29,3 +38,46 @@ export const valueToU256 = (value: bigint): U256 => [
   (value / u64Range ** 2n) % u64Range,
   (value / u64Range ** 3n) % u64Range,
 ]
+
+const parseReviveAddress = (address: SS58String | ReviveAddress | HexString) =>
+  typeof address === "string"
+    ? address.startsWith("0x")
+      ? Binary.fromHex(address)
+      : ss58ToEthereum(address)
+    : address
+
+// Ported from https://github.com/paritytech/polkadot-sdk/blob/c5ae50f86b8b727428eb86d9b1027a8f56fee19d/substrate/frame/revive/src/address.rs#L233
+export const getDeploymentAddressWithNonce = (
+  deployer: SS58String | ReviveAddress | HexString,
+  nonce: number,
+) => {
+  const addr = parseReviveAddress(deployer)
+  const data = RLP.encode([addr.asBytes(), nonce])
+  const bytes = keccak_256(data).slice(12)
+  return Binary.fromBytes(bytes)
+}
+
+// Ported from https://github.com/paritytech/polkadot-sdk/blob/c5ae50f86b8b727428eb86d9b1027a8f56fee19d/substrate/frame/revive/src/address.rs#L242
+export const getDeploymentAddressWithSalt = (
+  deployer: SS58String | ReviveAddress | HexString,
+  deploymentHash: HexString,
+  salt: HexString | FixedSizeBinary<32>,
+) => {
+  const addr = parseReviveAddress(deployer)
+  const saltBin = typeof salt === "string" ? Binary.fromHex(salt) : salt
+  const bytes = keccak_256(
+    mergeUint8(
+      new Uint8Array([0xff]),
+      addr.asBytes(),
+      saltBin.asBytes(),
+      Binary.fromHex(deploymentHash).asBytes(),
+    ),
+  ).slice(12)
+
+  return Binary.fromBytes(bytes)
+}
+
+export const getDeploymentHash = (code: Binary, inputData: Binary): HexString =>
+  Binary.fromBytes(
+    keccak_256(mergeUint8(code.asBytes(), inputData.asBytes())),
+  ).asHex()